backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydub==0.25.1
moviepy==1.0.3
Pillow==10.2.0
numpy==1.26.4
accelerate==0.30.0
diffusers==0.27.2
transformers==4.41.2
safetensors==0.4.2
# torch is environment-specific; install a CUDA build if available
# CPU-only (slow):
# torch==2.2.2+cpu torchvision==0.17.2+cpu torchaudio==2.2.2+cpu --index-url https://download.pytorch.org/whl/cpu
